backend/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Agno AI Framework
agno==1.4.5
openai==1.82.0

# Supabase
supabase==2.12.0

# Redis
redis==5.2.1

# HTTP Client
httpx==0.28.1

# Google Trends
pytrends==4.9.3

# Data Processing
pandas==2.2.3

# PDF Generation
reportlab==4.2.5
weasyprint==63.1

# Email
resend==2.7.0

# Celery (task queue)
celery[redis]==5.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0

# Auth
python-jose[cryptography]==3.3.0

# Rate Limiting
slowapi==0.1.9

# File uploads
python-multipart==0.0.20
